(ARM Use platform)change Makefile for arm platform using but Not Compile pass because of timerfd

+++ controller/lib/native_proccess/pipe.c
@@ -59,7 +59,7 @@ uint16_t read_pipe_tx( void *buf, uint16_t read_len )
 
 int write_pipe_tx(const void *buf, uint16_t buf_len )
 {
-	uint16_t pipe_len = 0;
+	int pipe_len = 0;
 	
 	pipe_len = write( fd[WRITE_PIPE_FD], buf,  buf_len );
 	if( pipe_len != -1 )
